Add support for running an individual test
Motivation:

Sometimes, typically when debugging a specific problem, it is useful to
run just a single test from the test-suite.  This allows for more
extensive debugging to be enabled.

Modification:

Add an optional second command-line argument to 'run-testsuite.sh' that
limits the testing to a specific test.

When only a single command-line argument is provided then the tests are
run as before.

Result:

It is now easy to run a single test from the suite.