Fix typo in the README document about read access vs read-only access
Motivation:

Currently, the documentation contains an oxymoron.  It says that the
storage system should be configured so that it grants:

> Read-only access [...] to all members of the WLCG VO

Not the use of the word "only" in read-only.

This means that "all members of the WLCG VO" should not be authorised to
write into the storage.

This is then immediately contradicted where the documentation describes
which users should be granted write access to different parts of the
namespace.

Modification:

Change "read-only access" to "read access".

Result:

The documentation is no longer self-contradictory.